# openapi.yaml
openapi: 3.0.0
info:
  title: Planter API
  description: API for the Planter application
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
...
  /notifications:
    get:
      tags:
        - Notifications
      summary: Get user notifications
      description: Get all notifications for the authenticated user
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
          description: Number of items per page
      responses:
        '200':
          description: List of notifications
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /notifications/{notificationId}/read:
    post:
      tags:
        - Notifications
      summary: Mark notification as read
      description: Mark a notification as read
      security:
        - bearerAuth: []
      parameters:
        - name: notificationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Notification marked as read
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Notification not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CareInstructions:
      type: object
      properties:
        wateringFrequency:
          type: integer
          description: Watering frequency in days
        sunlight:
          type: string
          enum:
            - LOW
            - MEDIUM
            - HIGH
        temperature:
          type: object
          properties:
            min:
              type: integer
            max:
              type: integer
        humidity:
          type: string
          enum:
            - LOW
            - MEDIUM
            - HIGH
        soilType:
          type: string
        fertilizerFrequency:
          type: integer
          description: Fertilizer frequency in days
        additionalNotes:
          type: string
    Plant:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        scientificName:
          type: string
        description:
          type: string
        imageUrl:
          type: string
        careInstructions:
          $ref: '#/components/schemas/CareInstructions'
        price:
          type: number
          nullable: true
        shopId:
          type: string
          format: uuid
          nullable: true
        isFavorite:
          type: boolean
        location:
          type: string
          nullable: true
        lastWatered:
          type: string
          format: date-time
          nullable: true
        nextWatering:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        error:
          type: string
    NotificationResponse:
      type: object
      properties:
        notifications:
          type: array
          items:
            $ref: '#/components/schemas/Notification'
        total:
          type: integer
          description: Total number of notifications
    Notification:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        plantId:
          type: string
          format: uuid
        type:
          type: string
          enum:
            - WATERING
        message:
          type: string
        isRead:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        plant:
          $ref: '#/components/schemas/Plant'